Validations

+++ app/Http/Controllers/ProductController.php
@@ -3,10 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Product;
-use Illuminate\Http\Request;
 use App\Http\Requests\ProductRequest;
-use App\Models\Comment;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -29,16 +28,20 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
-        $product = new Product();
-        $product->name = $request->get('name');
-        $product->price = $request->get('price');
-        $product->description = $request->get('description');
-        $product->amount = $request->get('amount');
-        $product->image_url = $request->get('image_url');
-        $product->price_descount = $request->get('price_descount');
-        $product->tag = $request->get('tag');
-        $product->save();
-        return response()->json($product, 201);
+        if (Gate::allows('isAdmin')) {
+            $product = new Product();
+            $product->name = $request->get('name');
+            $product->price = $request->get('price');
+            $product->description = $request->get('description');
+            $product->amount = $request->get('amount');
+            $product->image_url = $request->get('image_url');
+            $product->price_descount = $request->get('price_descount');
+            $product->tag = $request->get('tag');
+            $product->save();
+            return response()->json($product, 201);
+        } else {
+            return response()->json(['error' => 'No tiene permisos para hacer esta accion'], 401);
+        }
     }
 
     /**
@@ -49,14 +52,8 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
-        //$comment = Comment::get();
-        //$comments_of_products = $product->comments;
-        //$comments_of_product = DB::select("SELECT user_id from comments where product_id = :id", ['id'=> $product->id]);
-        //$user_name = DB::select("SELECT name, surname FROM users WHERE id = :id", ['id' => $comments_of_products->product->user_id]);
         $comments_with_names = DB::select("SELECT comments.*, users.name, users.surname FROM users INNER JOIN comments ON users.id = comments.user_id");
-        //$user_name = DB::table('users')->where('id', $comments_of_product->user_id)->value('name');
-        //$comments_of_products = $product->comments;
-        return response()->json(['product'=>$product, 'comments'=>$comments_with_names], 200);
+        return response()->json(['product' => $product, 'comments' => $comments_with_names], 200);
     }
 
     /**
@@ -68,39 +65,15 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
-        $product->name = $request->get('name');
-        $product->price = $request->get('price');
-        $product->description = $request->get('description');
-        $product->amount = $request->get('amount');
-        $product->image_url = $request->get('image_url');
-        $product->tag = $request->get('tag');
-        $product->save();
-        return response()->json($product,200);
-    }
-
-    public function addAmount(Product $product)
-    {
-        if($product->amount > 0){
-            $product->amount =  $product->amount+1;
+            $product->name = $request->get('name');
+            $product->price = $request->get('price');
+            $product->description = $request->get('description');
+            $product->amount = $request->get('amount');
+            $product->image_url = $request->get('image_url');
+            $product->tag = $request->get('tag');
             $product->save();
-            //return response()->json($product,200);
-        } else {
-            //return response()->json(['error'=> 'Ya no hay mas producto!'],200);
-        }
-
+            return response()->json($product, 200);
     }
-    /*
-    public function reduceAmount(Request $request)
-    {
-        if($product->amount > 0){
-            $product->amount =  $product->amount-1;
-            $product->save();
-            //return response()->json($product,200);
-        } else {
-            //return response()->json(['error'=> 'Ya no hay mas producto!'],200);
-        }
-    }
-    */
     /**
      * Remove the specified resource from storage.
      *
@@ -109,7 +82,11 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        $product->delete();
-        return response()->json(null, 204);
+        if (Gate::allows('isAdmin')) {
+            $product->delete();
+            return response()->json(null, 204);
+        } else {
+            return response()->json(['error' => 'No tiene permisos para hacer esta accion'], 401);
+        }
     }
 }
